wip: ed25519 signer
