+++ src/signer.ts
@@ -1,23 +1,14 @@
 import Account from './types/account.js';
 import { Transaction } from './transaction.js';
 import { encodeUnsignedSimulateTransaction } from './signedTransaction.js';
-import { LogicSigAccount, MSIG_PROGRAM_TAG, PROGRAM_TAG } from './logicsig.js';
+import { LogicSigAccount } from './logicsig.js';
 import { signLogicSigTransactionObject } from './signing.js';
-import { addressFromMultisigPreImg, MultisigMetadata } from './multisig.js';
+import { MultisigMetadata } from './multisig.js';
 import {
   signMultisigTransaction,
   mergeMultisigTransactions,
 } from './multisigSigning.js';
-import {
-  Address,
-  EncodedMultisig,
-  EncodedSubsig,
-  encodeMsgpack,
-  LogicSig,
-  SIGN_BYTES_PREFIX,
-  SignedTransaction,
-} from './main.js';
-import { arrayEqual, concatArrays } from './utils/utils.js';
+import { Address, EncodedMultisig, LogicSig } from './main.js';
 
 export interface Addressable {
   address: Address;
@@ -134,92 +125,6 @@ export function makeMultiSigAccountTransactionSigner(
     }
 
     return Promise.resolve(signed);
-  };
-}
-
-export type RawSigner = (bytesToSign: Uint8Array) => Promise<Uint8Array>;
-
-export function addressWithSignersFromRawSigner(
-  authAddress: Address,
-  rawSigner: RawSigner,
-  sendingAddress: Address = authAddress
-): AddressWithSigners {
-  const txnSigner: TransactionSigner = async (
-    txnGroup: Transaction[],
-    indexesToSign: number[]
-  ) => {
-    const stxns: SignedTransaction[] = [];
-    for (const index of indexesToSign) {
-      const txn = txnGroup[index];
-      // eslint-disable-next-line no-await-in-loop
-      const sig = await rawSigner(txn.bytesToSign());
-      const stxn = new SignedTransaction({
-        txn,
-        sig,
-        sgnr: txn.sender.equals(sendingAddress) ? undefined : authAddress,
-      });
-
-      stxns.push(stxn);
-    }
-
-    return stxns.map((stxn) => encodeMsgpack(stxn));
-  };
-
-  const delegatedLsigSigner: DelegatedLsigSigner = async (lsig, msig) => {
-    if (msig) {
-      const pks = msig.addrs.map(
-        (a) => Address.fromString(a.toString()).publicKey
-      );
-
-      const multisigAddr = addressFromMultisigPreImg({
-        version: msig.version,
-        threshold: msig.threshold,
-        pks,
-      });
-
-      const toBeSigned = concatArrays(
-        MSIG_PROGRAM_TAG,
-        multisigAddr.publicKey,
-        lsig.logic
-      );
-
-      const subsigs: EncodedSubsig[] = pks.map((pk) => ({ pk }));
-      const ourSubsig = subsigs.find((subsig) =>
-        arrayEqual(subsig.pk, authAddress.publicKey)
-      );
-      if (ourSubsig === undefined) {
-        throw Error(
-          `DelegatedLsigSigner could not find ${authAddress} as a signer in the multisig`
-        );
-      }
-
-      const sig = await rawSigner(toBeSigned);
-      ourSubsig.s = sig;
-      return {
-        address: authAddress,
-        lmsig: {
-          v: msig.version,
-          thr: msig.threshold,
-          subsig: subsigs,
-        } satisfies EncodedMultisig,
-      };
-    }
-
-    const toBeSigned = concatArrays(PROGRAM_TAG, lsig.logic);
-    const sig = await rawSigner(toBeSigned);
-    return { address: authAddress, sig };
-  };
-
-  const mxBytesSigner: MxBytesSigner = async (bytes: Uint8Array) => {
-    const bytesToSign = concatArrays(SIGN_BYTES_PREFIX, bytes);
-    return rawSigner(bytesToSign);
-  };
-
-  return {
-    address: sendingAddress,
-    txnSigner,
-    delegatedLsigSigner,
-    mxBytesSigner,
   };
 }
